+++ src/main/java/com/solvd/page_rank/dao/mybatis/PagesToRankDAO.java
@@ -1,4 +1,4 @@
-package com.solvd.page_rank.dao.dao.mybatis;
+package com.solvd.page_rank.dao.mybatis;
 
 import com.solvd.page_rank.dao.IPagesToRankDAO;
 import com.solvd.page_rank.models.PagesToRank;
