+++ src/main/java/com/solvd/page_rank/dao/mybatis/UsersDAO.java
@@ -1,7 +1,6 @@
-package com.solvd.page_rank.dao.dao.mybatis;
+package com.solvd.page_rank.dao.mybatis;
 
 import com.solvd.page_rank.dao.IUsersDAO;
-import com.solvd.page_rank.dao.dao.mybatis.AbstractDAO;
 import com.solvd.page_rank.models.PagesToRank;
 import com.solvd.page_rank.models.Users;
