+++ src/main/java/com/solvd/page_rank/utils/user_console_interface/WorkWithSettingMenu.java
@@ -1,6 +1,6 @@
 package com.solvd.page_rank.utils.user_console_interface;
 
-import com.solvd.page_rank.dao.dao.mybatis.SettingsForAlgorythmDAO;
+import com.solvd.page_rank.dao.mybatis.SettingsForAlgorythmDAO;
 import com.solvd.page_rank.exceptions.WrongNumberException;
 import com.solvd.page_rank.models.SettingsForAlgorythm;
 import org.apache.logging.log4j.LogManager;
